🏪 Market Update 🏪
➕ Added a -execute command
➕ Added a -marketadd command
➕ Added a -marketview command
➕ Added a timeFormatter(1h, 30h, 42m)
➕ Added a -suicide command
➕ Added so that the -reset command removes a users market listings
➕ Added so you can buy auctions through the -marketview command

🛠️ Fixed an issue with the sell command not selling individual items
🛠️ Fixed so that the help menu disappeares after the timer is done
🛠️ Fixed the -updaterules command to work after the new port

🗃️ Increased the help timer by 10 seconds
🗃️ When a user rebirths they are not allowed to have any items on the market
🗃️ FIxup and refactoring of almost all code
🗃️ Decided to do a full commit of almost all files used by the bot
🗃️ Ported over to postgressql

& m.m

🐞 0 Known

+++ adminCommands/Utils/ahelp.js
@@ -19,44 +19,49 @@ function sendDefaultHelpCommand(msg, client, disc, cmds){
 			name: `${categoryEmojiDict[category]} ${category}`,
 			value: `\u200b`,
 			inline: true
-		})
+		});
 	}
+
 	const filter = (reaction, user) => {
 		if(user.id == msg.member.id){
 			let index = Object.values(categoryEmojiDict).indexOf(reaction.emoji.name);
 			if (index != -1){
-				reaction.message.reactions.removeAll().then(message => {
-					message.delete();
-				});
+				reaction.message.reactions.removeAll()
+					.then(message => {
+						message.delete();
+					});
 				return sendSpecificHelpCommand(msg, client, disc, Object.keys(categoryEmojiDict)[index], cmds);
 			}
 		}
-		return false
+		return false;
 	}
 
 	msg.channel.send(embed)
 		.then(message => {
 			for(var key in categoryEmojiDict){
-				message.react(categoryEmojiDict[key])
+				message.react(categoryEmojiDict[key]);
 			}
-			message.awaitReactions(filter, {time: 15000}).then(m => {
-				if(!m.deleted) m.delete()
-			});
+			message.awaitReactions(filter, {time: 25000})
+				.then(m => {
+					if(!message.deleted) message.delete();
+				});
 		})
 }
 
 function sendSpecificHelpCommand(msg, client, disc, category, cmds){
-	if (category == undefined) return
+	if (category == undefined) return;
 
-	const categoryIndex = client.adminCategoryList.map((category) => category.toLowerCase()).indexOf(category.toLowerCase())
-	if(categoryIndex == -1) return
+	const categoryIndex = client.adminCategoryList.map((category) => category.toLowerCase()).indexOf(category.toLowerCase());
+
+	if(categoryIndex == -1) return;
 
 	const embed = new disc.MessageEmbed()
 		.setAuthor(`${client.adminCategoryList[categoryIndex]}'s Command List`)
-		.setColor(`#41BDB8`)
+		.setColor(`#41BDB8`);
 
 	for(let cmd of cmds){
-		if(cmd.options.Category.toLowerCase() != client.adminCategoryList[categoryIndex].toLowerCase()) continue
+		if(cmd.options.Category.toLowerCase() != client.adminCategoryList[categoryIndex].toLowerCase()) continue;
+
 		embed.addFields({
 			name: `${cmd.name}`,
 			value: `${cmd.description}`,
@@ -71,12 +76,12 @@ function sendSpecificHelpCommand(msg, client, disc, category, cmds){
 			name: `Aliases`,
 			value: (cmd.alias.length != 0 ? cmd.alias : '\u200b'),
 			inline: true
-		})
+		});
 	}
 	
-	msg.channel.send(embed)
+	msg.channel.send(embed);
 	
-	return true
+	return true;
 }
 
 module.exports = {
@@ -88,12 +93,12 @@ module.exports = {
 	run : function(msg, client, disc, args){
 		try{
 			const cmds = client.adminCommands.array();	
-			if(sendSpecificHelpCommand(msg, client, disc, args[0], cmds)) return
+			if(sendSpecificHelpCommand(msg, client, disc, args[0], cmds)) return;
 			
-			sendDefaultHelpCommand(msg, client, disc, cmds)
+			sendDefaultHelpCommand(msg, client, disc, cmds);
 
         }catch(e){
-            client.eventEm.emit('CommandError', msg, this.name, args, e)
+            client.eventEm.emit('CommandError', msg, this.name, args, e);
         }
 	}
 }
